Split getTemplateVersionState to its own file
